requirements.txt
# Core
youtube-transcript-api>=0.6.1
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# NLP
nltk>=3.8.0

# UI
streamlit>=1.30.0

# Evaluation
ragas>=0.1.0
rouge-score>=0.1.2
bert-score>=0.3.13
datasets>=2.14.0

# Visualization & Data
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
